glib: gobject: gparamspecs: Fix implicit conversion
external/mainline-hw-deps/glib/gobject/gparamspecs.c:51:20: error: implicit conversion from 'int' to 'gint8' (aka 'signed char') changes value from 255 to -1 [-Werror,-Wconstant-conversion]
   51 |   cspec->maximum = CHAR_MAX;
      |                  ~ ^~~~~~~~
prebuilts/clang/host/linux-x86/clang-r563880c/lib/clang/21/include/limits.h:93:18: note: expanded from macro 'CHAR_MAX'
   93 | #define CHAR_MAX UCHAR_MAX
      |                  ^~~~~~~~~
prebuilts/clang/host/linux-x86/clang-r563880c/lib/clang/21/include/limits.h:58:37: note: expanded from macro 'UCHAR_MAX'
   58 | #define UCHAR_MAX (__SCHAR_MAX__*2  +1)
      |                    ~~~~~~~~~~~~~~~~~^~
1 error generated.

+++ gobject/gparamspecs.c
@@ -48,7 +48,7 @@ param_char_init (GParamSpec *pspec)
   GParamSpecChar *cspec = G_PARAM_SPEC_CHAR (pspec);
   
   cspec->minimum = CHAR_MIN;
-  cspec->maximum = CHAR_MAX;
+  cspec->maximum = (gint8)CHAR_MAX;
   cspec->default_value = 0;
 }
